:truck: Move MediaFormatChecker to util package
Relocate MediaFormatChecker class and update imports across affected files
