+++ vvd-extractor/src/main/kotlin/mikufan/cx/vvd/extractor/component/util/MediaFormatChecker.kt
@@ -1,4 +1,4 @@
-package mikufan.cx.vvd.extractor.component
+package mikufan.cx.vvd.extractor.component.util
 
 import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.databind.ObjectMapper
